requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
npx==0.1.2
numpy==1.26.4
optional-django==0.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2024.1
react==4.3.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
